Restore the engine tests
And make them pass.
With this, we have confidence that the core functionality of
the assistant engine is still there.
