+++ lib/dialogue-agent/dispatcher.js
@@ -51,6 +51,7 @@ module.exports = class Assistant extends events.EventEmitter {
         super();
 
         this._engine = engine;
+        this._outputs = new Set;
         this._conversations = new Map;
         this._lastConversation = null;
     }
@@ -86,8 +87,17 @@ module.exports = class Assistant extends events.EventEmitter {
         return result;
     }
 
+    addNotificationOutput(output) {
+        this._outputs.add(output);
+    }
+    removeNotificationOutput(output) {
+        this._outputs.delete(output);
+    }
+
     async notifyAll(...data) {
         const promises = [];
+        for (let out of this._outputs.values())
+            promises.push(out.notify(...data));
         for (let conv of this._conversations.values())
             promises.push(conv.notify(...data));
         await Promise.all(promises);
@@ -95,6 +105,8 @@ module.exports = class Assistant extends events.EventEmitter {
 
     async notifyErrorAll(...data) {
         const promises = [];
+        for (let out of this._outputs.values())
+            promises.push(out.notifyError(...data));
         for (let conv of this._conversations.values())
             promises.push(conv.notifyError(...data));
         await Promise.all(promises);
