+++ lib/dialogue-agent/execution/statement_executor.js
@@ -124,29 +124,36 @@ class InferenceStatementExecutor {
         return new Ast.Value.Object(result);
     }
 
+    async _iterateResults(app, into) {
+        let count = 0;
+        if (app === null)
+            return false;
+
+        for await (const value of app.mainOutput) {
+            if (count >= MORE_SIZE)
+                return true;
+
+            if (value instanceof Error) {
+                console.error(`FIXME: unhandled error from ThingTalk`, value);
+            } else {
+                const raw = [value.outputType, value.outputValue];
+                const mapped = await this._mapResult(value.outputType, value.outputValue);
+                into.push(new Ast.DialogueHistoryResultItem(null, mapped, raw));
+                count ++;
+            }
+        }
+
+        // if we get here, we iterated all results from the app, so we can stop
+        return false;
+    }
+
     async executeStatement(stmt) {
         assert(stmt instanceof Ast.Statement.Command || stmt instanceof Ast.Statement.Rule);
 
         const program = new Ast.Program(null, [], [], [stmt]);
         const app = await this._engine.createApp(program);
         const results = [];
-        let more = false;
-        await app.runCommand({
-            output: async (outputType, outputValue) => {
-                if (results.length >= MORE_SIZE) {
-                    more = true;
-                    return false;
-                }
-                const raw = [outputType, outputValue];
-                const mapped = await this._mapResult(outputType, outputValue);
-                results.push(new Ast.DialogueHistoryResultItem(null, mapped, raw));
-                return results.length < MORE_SIZE;
-            },
-
-            notifyError(error) {
-                console.error(`FIXME: unhandled error from ThingTalk`, error);
-            }
-        });
+        const more = await this._iterateResults(app, results);
 
         const resultList = new Ast.DialogueHistoryResultList(null, results.slice(0, PAGE_SIZE),
             new Ast.Value.Number(results.length), more);
