+++ lib/engine/apps/app_executor.js
@@ -10,12 +10,65 @@
 "use strict";
 
 const events = require('events');
+const AsyncQueue = require('consumer-queue');
 
 const ThingTalk = require('thingtalk');
 const AppGrammar = ThingTalk.Grammar;
 const AppCompiler = ThingTalk.Compiler;
 const RuleExecutor = require('./rule_executor');
 const { ChannelState } = require('../db/channel');
+
+class QueueOutputDelegate {
+    constructor() {
+        this._queue = new AsyncQueue();
+    }
+
+    [Symbol.asyncIterator]() {
+        return this;
+    }
+    next() {
+        return this._queue.pop();
+    }
+
+    done() {
+        this._queue.push({ done: true });
+    }
+    output(outputType, outputValue) {
+        this._queue.push({ done: false, value: { outputType, outputValue } });
+    }
+    notifyError(error) {
+        this._queue.push({ done: false, value: error });
+    }
+}
+
+class NotificationOutputDelegate {
+    constructor(app) {
+        this._app = app;
+        this._engine = app.engine;
+    }
+
+    done() {}
+
+    /**
+     * Report that the app had an error.
+     * @param {Error} error - the error that occurred.
+     * @package
+     */
+    notifyError(error) {
+        this._app.error = error;
+        return this._engine.assistant.notifyErrorAll(this._app.uniqueId, this._app.icon, error);
+    }
+
+    /**
+     * Report a new result from app.
+     * @param {string} outputType - the type of result.
+     * @param {any} outputValue - the actual result.
+     * @package
+     */
+    output(outputType, outputValue) {
+        return this._engine.assistant.notifyAll(this._app.uniqueId, this._app.icon, outputType, outputValue);
+    }
+}
 
 /**
  * The representation of a currently executing ThingTalk program.
@@ -98,20 +151,13 @@ class AppExecutor extends events.EventEmitter {
         this._finished = false;
 
         this._states = [];
+
+        this.mainOutput = new QueueOutputDelegate();
+        this._notificationOutput = new NotificationOutputDelegate(this);
     }
 
     get metadata() {
         return this._meta;
-    }
-
-    getConversation() {
-        var assistant = this.engine.platform.getCapability('assistant');
-        if (!assistant)
-            throw new Error('Assistant not supported');
-        if (this.conversation)
-            return assistant.getConversation(this.conversation);
-        else
-            return null;
     }
 
     _updateNameDescription(name, description) {
@@ -138,24 +184,8 @@ class AppExecutor extends events.EventEmitter {
     set error(e) {
         this._error = e;
     }
-    /**
-     * Report that the app had an error.
-     * @param {Error} error - the error that occurred.
-     * @package
-     */
-    notifyError(error) {
-        this._error = error;
-        return this._engine.assistant.notifyErrorAll(this.uniqueId, this.icon, error);
-    }
-
-    /**
-     * Report a new result from app.
-     * @param {string} outputType - the type of result.
-     * @param {any} outputValue - the actual result.
-     * @package
-     */
-    output(outputType, outputValue) {
-        return this._engine.assistant.notifyAll(this.uniqueId, this.icon, outputType, outputValue);
+    reportError(error) {
+        this._notificationOutput.notifyError(error);
     }
 
     /**
@@ -199,11 +229,11 @@ class AppExecutor extends events.EventEmitter {
         const compiled = await this.compiler.compileProgram(this._ast);
 
         if (compiled.command)
-            this.command = new RuleExecutor(this.engine, this, compiled.command);
+            this.command = new RuleExecutor(this.engine, this, compiled.command, this.mainOutput);
 
         this._finishedRules = new Set;
         for (let rule of compiled.rules) {
-            const executor = new RuleExecutor(this.engine, this, rule);
+            const executor = new RuleExecutor(this.engine, this, rule, this._notificationOutput);
             this.rules.push(executor);
             executor.on('finish', () => {
                 this._finishedRules.add(executor);
@@ -228,9 +258,8 @@ class AppExecutor extends events.EventEmitter {
      *
      * @async
      */
-    async runCommand(delegate) {
+    async runCommand() {
         if (this.command) {
-            this.command.setOutput(delegate);
             this.command.start();
             await this.command.waitFinished();
         }
