+++ lib/engine/apps/database.js
@@ -51,6 +51,8 @@ class AppDatabase extends events.EventEmitter {
 
     async _doStartApp(app, isNewApp) {
         try {
+            if (isNewApp)
+                await app.runCommand();
             // only start and save into db apps that actually have some rules
             if (app.rules.length > 0) {
                 this._enableApp(app);
@@ -129,15 +131,9 @@ class AppDatabase extends events.EventEmitter {
 
     _addAppInternal(app, uniqueId) {
         if (app.uniqueId === undefined) {
-            // HACK: we accept null as well as undefined because undefined
-            // is lost by the RPC layer in the cloud case
-            if (uniqueId === undefined || uniqueId === null)
-                app.uniqueId = 'uuid-' + uuid.v4();
-            else
-                app.uniqueId = uniqueId;
+            app.uniqueId = uniqueId;
         } else {
-            if ((uniqueId !== undefined && uniqueId !== null) &&
-                app.uniqueId !== uniqueId)
+            if (uniqueId !== undefined && app.uniqueId !== uniqueId)
                 throw new Error('App unique id is different from stored value');
         }
