+++ lib/engine/apps/exec_wrapper.js
@@ -35,21 +35,21 @@ function extendParams(output, input) {
  * @package
  */
 class ExecWrapper extends ExecEnvironment {
-    constructor(engine, app) {
+    constructor(engine, app, output) {
         super(engine.platform.locale, engine.platform.timezone, engine.schemas,
               engine.platform.getCapability('gettext'));
         this.engine = engine;
         this.app = app;
         this._programId = new ThingTalk.Entity(this.app.uniqueId, null);
-
-        // by default, output to the app itself (which will broadcast on all conversations)
-        // when running interactively, this will be overridden to output to a specific
-        // StatementExecutor
-        this.output = app;
+        this._outputDelegate = output;
 
         this._trigger = null;
 
         this._execCache = [];
+    }
+
+    setOutput(delegate) {
+        this._outputDelegate = delegate;
     }
 
     get program_id() {
@@ -186,11 +186,11 @@ class ExecWrapper extends ExecEnvironment {
         return this.app.writeState(stateId, state);
     }
 
-    invokeAction(kind, attrs, fname, params) {
+    async invokeAction(kind, attrs, fname, params) {
         const devices = this._getDevices(kind, attrs);
         const js_function_name = 'do_' + fname;
         return Promise.all(devices.map(async (d) => {
-            return d[js_function_name](params, this);
+            return [d.kind + ':' + fname, await d[js_function_name](params, this)];
         }));
     }
 
@@ -209,11 +209,11 @@ class ExecWrapper extends ExecEnvironment {
 
         console.error(message, error);
         this.app.error = error;
-        return this.output.notifyError(error);
+        return this._outputDelegate.notifyError(error);
     }
 
     output(outputType, outputValues) {
-        return this.output.output(outputType, outputValues);
+        return this._outputDelegate.output(outputType, outputValues);
     }
 }
 module.exports = ExecWrapper;
