+++ lib/engine/apps/rule_executor.js
@@ -14,12 +14,13 @@ const events = require('events');
 const ExecWrapper = require('./exec_wrapper');
 
 module.exports = class RuleExecutor extends events.EventEmitter {
-    constructor(engine, app, compiled) {
+    constructor(engine, app, compiled, output) {
         super();
         this.engine = engine;
         this.app = app;
 
-        this._env = new ExecWrapper(engine, app);
+        this._output = output;
+        this._env = new ExecWrapper(engine, app, output);
         this._tt = compiled;
 
         // create an early promise so we can call waitFinished() before start()
@@ -27,10 +28,6 @@ module.exports = class RuleExecutor extends events.EventEmitter {
         this._finishPromise = new Promise((resolve, reject) => {
             this._finished = { resolve, reject };
         });
-    }
-
-    setOutput(output) {
-        this._env.output = output;
     }
 
     async _ruleThread() {
@@ -41,6 +38,7 @@ module.exports = class RuleExecutor extends events.EventEmitter {
         }
 
         this.emit('finish');
+        this._output.done();
         this._finished.resolve();
     }
