+++ lib/engine/devices/builtins/thingengine.builtin.js
@@ -109,10 +109,18 @@ module.exports = class MiscellaneousDevice extends Tp.BaseDevice {
     do_debug_log(args) {
         console.log('DEBUG:', args.message);
     }
-    do_say(args, env) {
-        return env.say(args.message);
+    do_say(args) {
+        return { message_output: args.message };
     }
+    do_open_url({ url }) {
+        let cap = this.engine.platform.getCapability('app-launcher');
+        if (!cap)
+            throw new Error(this.engine._("Opening files is not implemented in this Almond"));
+        return cap.launchURL(String(url));
+    }
+
     do_configure(args, env) {
+        /*
         var conversation = env.app.getConversation();
         if (!conversation)
             throw new Error(this.engine._("User not available"));
@@ -120,8 +128,11 @@ module.exports = class MiscellaneousDevice extends Tp.BaseDevice {
         Promise.resolve(conversation.interactiveConfigure(String(args.device))).catch((err) => {
             // the error has already been logged by Almond, and the user has been informed
         });
+        */
+        throw new Error('not implemented');
     }
     do_discover(args, env) {
+        /*
         var conversation = env.app.getConversation();
         if (!conversation)
             throw new Error(this.engine._("User not available"));
@@ -129,11 +140,7 @@ module.exports = class MiscellaneousDevice extends Tp.BaseDevice {
         Promise.resolve(conversation.interactiveConfigure(null)).catch((err) => {
             // the error has already been logged by Almond, and the user has been informed
         });
-    }
-    do_open_url({ url }) {
-        let cap = this.engine.platform.getCapability('app-launcher');
-        if (!cap)
-            throw new Error(this.engine._("Opening files is not implemented in this Almond"));
-        return cap.launchURL(String(url));
+        */
+        throw new Error('not implemented');
     }
 };
