--- test/engine/everything.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-
-set -e
-set -x
-srcdir=`dirname $0`
-node $srcdir/unit
-node $srcdir/functional
