+++ test/engine/index.js
@@ -16,37 +16,14 @@ process.on('unhandledRejection', (up) => { throw up; });
 // require(thingpedia) to initialize the polyfill
 require('thingpedia');
 
-const assert = require('assert');
-
 const Engine = require('../../lib/engine');
-
-class MockAssistant {
-    constructor() {
-    }
-
-    _setConversation(conv) {
-        this._conv = conv;
-    }
-
-    getConversation(conv) {
-        assert(conv === 'mock' || conv === undefined);
-        return this._conv;
-    }
-
-    notifyAll(...data) {
-        this._conv.notify(...data);
-    }
-    notifyErrorAll(...data) {
-        this._conv.notifyError(...data);
-    }
-}
 
 const THINGPEDIA_URL = 'https://almond-dev.stanford.edu/thingpedia';
 const THINGENGINE_URL = 'https://almond-dev.stanford.edu';
 
 async function runTests(engine, limitTo) {
     try {
-        for (let x of ['devices', 'apps', 'database', 'http_client', 'builtins', 'permissions', 'remote', 'messaging', 'cloud_sync']) {
+        for (let x of ['devices', 'apps', 'database', 'http_client', 'builtins', 'cloud_sync']) {
             if (limitTo !== undefined && x !== limitTo)
                 continue;
             console.log(`Running ${x} tests`);
@@ -62,13 +39,12 @@ async function runTests(engine, limitTo) {
 
 async function main() {
     const platform = require('./platform').newInstance();
-    platform.setAssistant(new MockAssistant());
-
     const engine = new Engine(platform, {
         thingpediaUrl: THINGPEDIA_URL,
         cloudSyncUrl: THINGENGINE_URL
     });
     await engine.open();
+    await engine.assistant.openConversation('mock');
 
     runTests(engine, process.argv[2]);
     await engine.run();
