+++ test/engine/test-classes/test_platform.js
