+++ test/unit/index.js
@@ -29,16 +29,22 @@ async function do_test(array) {
 
 // test lib scripts
 do_test([
-    ('./test_misc'),
-    ('./test_random'),
-    ('./test_stream_utils'),
-    ('./test_requoting'),
-    ('./test_sentence_generator'),
+    ('./test_array_set'),
     ('./test_augment'),
+    ('./test_engine_protocol'),
+    ('./test_entities'),
+    ('./test_helpers'),
     ('./test_i18n_chinese'),
+    ('./test_ip_address'),
+    ('./test_linked_list'),
+    ('./test_misc'),
     ('./test_new_tokenizer_en'),
     ('./test_new_tokenizer_it'),
     ('./test_new_tokenizer_zh'),
-    ('./test_helpers'),
-    ('./test_entities'),
+    ('./test_random'),
+    ('./test_requoting'),
+    ('./test_sentence_generator'),
+    ('./test_stream_utils'),
+    ('./test_timed_reference'),
+    ('./test_timers'),
 ]);
